move implementation of org subcommand to org package
